Oprava kompilacni chyby Prepsal jsem v CodeGenerator stare nazvy chyb, aby slo vse zkompilovat Opravil jsem "always true" warning v BottomUp zmenou datoveho typu pro pravidlo v datove strukture.

+++ Source/PrecedentTable.h
@@ -24,7 +24,7 @@ typedef struct IdxTerminalPair {
 	Terminal type;
 	Terminal* funcParams;
 	const char* funcName;
-	size_t rule;
+	int rule;
 	int error;
 }IdxTerminalPair;
